requirements.txt
alembic==1.16.4
annotated-types==0.7.0
asgiref==3.9.1
astroid==3.3.10
asttokens==3.0.0
bcrypt==4.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
bs4==0.0.2
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
decorator==5.2.1
dill==0.4.0
distlib==0.4.0
Django==5.2.4
django-cors-headers==4.9.0
django-debug-toolbar==6.0.0
django-ninja==1.4.3
django-phonenumber-field==8.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
executing==2.2.0
filelock==3.18.0
Flask==3.1.1
Flask-Bcrypt==1.0.1
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
gunicorn==23.0.0
idna==3.10
ipython==9.4.0
ipython_pygments_lexers==1.1.1
isort==6.0.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
lxml==6.0.0
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mysqlclient==2.2.7
openpyxl==3.1.5
packaging==25.0
parso==0.8.4
phonenumbers==9.0.15
pipenv==2025.0.4
platformdirs==4.3.8
prompt_toolkit==3.0.51
pure_eval==0.2.3
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pylint==3.3.7
python-decouple==3.8
python-dotenv==1.1.1
requests==2.32.4
setuptools==80.9.0
soupsieve==2.7
SQLAlchemy==2.0.41
sqlparse==0.5.3
stack-data==0.6.3
tomlkit==0.13.3
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
virtualenv==20.32.0
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
